Separate GenericObjectDeserializer
