+++ Assets/VRM/UniJSON/Scripts/JsonSchemaValidator/GenericObjectDeserializer.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 2f3c58dc82ad74549916f824156baad6
+timeCreated: 1550087061
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
